registration validate 1

+++ src/app/views/pages/register/register.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import{ RegisterUser } from './registerUser';
 import { RegisterService } from './register.service';
 import { Router } from '@angular/router';
+import {FormGroup,FormControl,Validators} from '@angular/forms';
 @Component({
   selector: 'app-register',
   templateUrl: './register.component.html',
@@ -9,13 +10,14 @@ import { Router } from '@angular/router';
   providers: [RegisterService]
 })
 export class RegisterComponent implements OnInit {
-
-
+   
+isfilled:boolean =false;
+ispasswordMatch:boolean =true;
 registeruser:RegisterUser ={
 
   firstname: "",
   lastname:  "",
-  usernaem: "",
+  username: "",
   email:"",
   password: "",
   confirmpassword: "",
@@ -26,32 +28,47 @@ registeruser:RegisterUser ={
 
 isRegistrationFailed: boolean = false;
 
-  constructor( private registerservice:RegisterService, private router:Router) { }
+  constructor( private registerservice:RegisterService, private router:Router) { 
+
+  }
 
   ngOnInit(){
 
     if(localStorage.getItem("user")) {
-      this.router.navigate(['/dashbord']);
+      //this.router.navigate(['dashboard']);
     }
   }
-
+  
 
 //register user
 
 register():void{
   console.log("registration success");
+  if(!this.registeruser.firstname || !this.registeruser.lastname || !this.registeruser.username || !this.registeruser.email|| !this.registeruser.password || !this.registeruser.confirmpassword){
+        this.isfilled = true;
+        console.log('not feeld'+this.isfilled);
+
+  
+  
+  
+}
+
+if(!(this.registeruser.password == this.registeruser.confirmpassword) ){
+  this.ispasswordMatch =false;
+
+}else {
   this.registerservice.register(this.registeruser)
   .then(data => {
     console.log("registration success");
     localStorage.setItem("user", JSON.stringify(data));
-    this.router.navigate(['/dashboard']);
+    this.router.navigate(['/pages/login']);
   })
   .catch(error => {
     console.log(error);
     this.isRegistrationFailed = true;
   });
 
-
+}
 
 }
